Refactored namespaces and added build props
