+++ Directory.Build.props
@@ -1,0 +1,7 @@
+<Project>
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+        <LangVersion>9.0</LangVersion>
+        <NeutralLanguage>en</NeutralLanguage>
+    </PropertyGroup>
+</Project>
