--- Gomoku.Contracts/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Gomoku.Contracts
-{
-    public class Class1
-    {
-    }
-}
