+++ Gomoku.Storage/DatabaseContext.cs
@@ -1,10 +1,5 @@
 ﻿using Gomoku.Storage.Models;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Gomoku.Storage
 {
@@ -12,7 +7,12 @@ namespace Gomoku.Storage
     {
         public DatabaseContext(DbContextOptions<DatabaseContext> optionsBuilder) : base(optionsBuilder) { }
 
-        public DbSet<Player> Players { get; set; }
+        public DatabaseContext()
+        {
+            
+        }
+
+        public DbSet<Player> Players { get; set; } 
         public DbSet<GameTurn> GameTurns { get; set; }
     }
 }
