+++ Gomoku.WebApi/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DatabaseContext": "Server=localhost;Port=5432;database=Gomoku;uid=postgres;pwd=password;"
+    "DatabaseContext": "Server=localhost;Port=5432;database=Gomoku;uid=postgres;password=secret"
   }
 }
