+++ Gomoku/Services/StatisticsService.cs
@@ -1,9 +1,8 @@
-﻿using Gomoku.Contracts.Models;
+﻿using Gomoku.Contracts.Responses;
 using Gomoku.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
-using System.Text;
 using System.Text.Json;
 
 namespace Gomoku.Services
@@ -14,13 +13,14 @@ namespace Gomoku.Services
 
         public StatisticsService()
         {
-            _httpClient = new HttpClient() { BaseAddress = new Uri("https://localhost:5001") };
+            _httpClient = new HttpClient() { BaseAddress = new Uri("https://localhost:5001")};
         }
 
         public ICollection<PlayerResponse> GetPlayersData()
         {
             var response = _httpClient.GetAsync("/players").Result;
             var players = JsonSerializer.Deserialize<ICollection<PlayerResponse>>(response.Content.ReadAsStringAsync().Result);
+
             return players;
         }
     }
